WIP. Доработка. - Загрузка плиток в выделнном потоке. - Добавлена точность для индекса, который теперь   имеет разрядность 64 бита со знаком. - Добавлена точность для вычислений, теперь   используется double вместо float. - Увеличен размер окна.

+++ headers/tiletools.h
@@ -53,7 +53,7 @@ float deg_to_rad(float v);
 /// @param ty Индекс плитки по оси Y.
 /// @param tz Уровень детализации плитки.
 /// @return Индекс одномерного массива.
-uint32_t tile_to_index(float tx, float ty, float tz);
+int64_t tile_to_index(float tx, float ty, float tz);
 
 
 /// @brief Преобразует долготу и широту в индекс плитки .
